requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-crispy-forms==1.14.0
django-filter==22.1
Pillow==9.2.0
sqlparse==0.4.2
